Report backup size in bytes instead of MB (#131028)

+++ homeassistant/components/backup/models.py
@@ -11,7 +11,7 @@ class BaseBackup:
     date: str
     name: str
     protected: bool
-    size: float
+    size: int
 
     def as_dict(self) -> dict:
         """Return a dict representation of this backup."""
